Avoid reading the whole file.
Some DLL files are really big (hundreds of megabytes). Loading them as a whole is both slow and useless.
Also, in some cases it may cause system hangs in DosRead due to bugs in the OS/2 Kernel and/or IFS
(see https://github.com/bitwiseworks/libc/issues/84#issuecomment-664380236 for details).

The new code reads just a few hundred of bytes of the DLL when processing it.